Move util function into its own file
