+++ src/modlunky2/sprites/base_classes/base_sprite_loader.py
@@ -5,31 +5,13 @@ from typing import Callable, Dict, Optional, Union
 
 from PIL import Image
 
-from .types import chunk_map_type, image_crop_tuple
+from .types import chunk_map_type
 
 _DEFAULT_BASE_PATH = Path(
     r"C:\Program Files (x86)\Steam\steamapps\common\Spelunky 2\Mods\Extracted"
 )
 
 _CACHED_NONE_SENTINEL = object()
-
-
-def chunks_from_animation(
-    base_name: str, start_chunk: image_crop_tuple, num_chunks: int
-):
-    chunk_width = start_chunk[2] - start_chunk[0]
-
-    chunks = {}
-    for i in range(0, num_chunks):
-        coords = (
-            start_chunk[0] + i * chunk_width,
-            start_chunk[1],
-            start_chunk[2] + i * chunk_width,
-            start_chunk[3],
-        )
-        chunks["{}{}".format(base_name, i + 1)] = coords
-
-    return chunks
 
 
 def _cache_img_not_class(f):
