+++ src/modlunky2/sprites/character.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
-from .base_classes import BaseSpriteLoader, chunks_from_animation
+from .base_classes import BaseSpriteLoader
+from .util import chunks_from_animation
 
 
 class CharacterSheet(BaseSpriteLoader):
